only existent users can be deleted

+++ fastify_app/fastify/controllers/controls/delete.controller.js
@@ -12,11 +12,17 @@ export async function deleteUser(request, reply) {
 			console.error("Database is not initialized");
 			return reply.send({ error: "Database connection error" });
 		}
-
-		const stmt = db.prepare("DELETE FROM userTable WHERE username = ? AND password = ?");
-		const result = stmt.run(username, password);
-		console.log("User deleted: " + username);
-		return reply.send({ message: `User successfully deleted: ${username}` });
+		const stmt = db.prepare('SELECT * FROM userTable WHERE username = ? AND password = ?');
+		const user = stmt.get(username, password);
+		if (user)
+		{
+			const stmt = db.prepare("DELETE FROM userTable WHERE username = ? AND password = ?");
+			const result = stmt.run(username, password);
+			console.log("User deleted: " + username);
+			return reply.send({ message: `User successfully deleted: ${username}` });
+		}
+		else
+			reply.send({ error: "Invalid credentials / non-existent user" });
 	}
 	catch (err) {
 		console.log(err);
